+++ crates/common/src/cache/tests.rs
@@ -5143,7 +5143,7 @@ impl CacheDatabaseAdapter for SnapshotBlobTestDatabase {
         Ok(AHashMap::new())
     }
 
-    fn load_index_order_position(&self) -> anyhow::Result<AHashMap<ClientOrderId, Position>> {
+    fn load_index_order_position(&self) -> anyhow::Result<AHashMap<ClientOrderId, PositionId>> {
         Ok(AHashMap::new())
     }